+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## [4.5.3] - 2026-03-25
+### Added
+- deploy separate virtual tokens for USD, EUR, BTC
+- deploy `VirtualTokenPrice` contract
+
 ## [4.5.2] - 2026-03-25
 ### Added
 - add ORACLE Dao for Injective